+++ ast/ast.go
@@ -50,6 +50,11 @@ type IntegerLiteral struct {
 	Value int64
 }
 
+type Boolean struct {
+	Token token.Token
+	Value bool
+}
+
 type PrefixExpression struct {
 	Token    token.Token // prefix token; !, -
 	Operator string
@@ -96,6 +101,10 @@ func (es *ExpressionStatement) TokenLiteral() string { return es.Token.Literal }
 func (il *IntegerLiteral) expressionNode()      {}
 func (il *IntegerLiteral) TokenLiteral() string { return il.Token.Literal }
 func (il *IntegerLiteral) String() string       { return il.Token.Literal }
+
+func (b *Boolean) expressionNode()      {}
+func (b *Boolean) TokenLiteral() string { return b.Token.Literal }
+func (b *Boolean) String() string       { return b.Token.Literal }
 
 func (pe *PrefixExpression) expressionNode()      {}
 func (pe *PrefixExpression) TokenLiteral() string { return pe.Token.Literal }